+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.8...@wundergraph/nextjs@0.9.9) (2023-03-28)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
 ## [0.9.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.9.7...@wundergraph/nextjs@0.9.8) (2023-03-27)
 
 **Note:** Version bump only for package @wundergraph/nextjs